+++ src/internal/utils/CommonUtils.ts
